+++ cognite/client/_api/assets.py
@@ -18,8 +18,11 @@ class AssetsAPI(APIClient):
         chunk_size: int = None,
         name: str = None,
         parent_ids: List[int] = None,
+        parent_external_ids: List[str] = None,
         root_ids: List[int] = None,
         root_external_ids: List[str] = None,
+        asset_subtree_ids: List[int] = None,
+        asset_subtree_external_ids: List[str] = None,
         metadata: Dict[str, str] = None,
         source: str = None,
         created_time: Union[Dict[str, Any], TimestampRange] = None,
@@ -37,9 +40,12 @@ class AssetsAPI(APIClient):
         Args:
             chunk_size (int, optional): Number of assets to return in each chunk. Defaults to yielding one asset a time.
             name (str): Name of asset. Often referred to as tag.
-            parent_ids (List[int]): No description.
+            parent_ids (List[int]): Return only the direct descendants of the specified assets.
+            parent_external_ids (List[str]): Return only the direct descendants of the specified assets.
             root_ids (List[int], optional): List of root ids ids to filter on.
             root_external_ids (List[str], optional): List of root external ids to filter on.
+            asset_subtree_ids (List[int]): List of asset subtrees ids to filter on.
+            asset_subtree_external_ids (List[str]): List of asset subtrees external ids to filter on.
             metadata (Dict[str, str]): Custom, application specific metadata. String key -> String value
             source (str): The source of this asset
             created_time (Union[Dict[str, Any], TimestampRange]): Range between two timestamps
@@ -58,11 +64,15 @@ class AssetsAPI(APIClient):
         # dict option for backward compatibility
         if (root_ids and not isinstance(root_ids[0], dict)) or root_external_ids:
             root_ids = self._process_ids(root_ids, root_external_ids, wrap_ids=True)
+        if asset_subtree_ids or asset_subtree_external_ids:
+            asset_subtree_ids = self._process_ids(asset_subtree_ids, asset_subtree_external_ids, wrap_ids=True)
 
         filter = AssetFilter(
             name=name,
             parent_ids=parent_ids,
+            parent_external_ids=parent_external_ids,
             root_ids=root_ids,
+            asset_subtree_ids=asset_subtree_ids,
             metadata=metadata,
             source=source,
             created_time=created_time,
@@ -115,12 +125,18 @@ class AssetsAPI(APIClient):
         utils._auxiliary.assert_exactly_one_of_id_or_external_id(id, external_id)
         return self._retrieve_multiple(ids=id, external_ids=external_id, wrap_ids=True)
 
-    def retrieve_multiple(self, ids: Optional[List[int]] = None, external_ids: Optional[List[str]] = None) -> AssetList:
+    def retrieve_multiple(
+        self,
+        ids: Optional[List[int]] = None,
+        external_ids: Optional[List[str]] = None,
+        ignore_unknown_ids: bool = False,
+    ) -> AssetList:
         """`Retrieve multiple assets by id. <https://docs.cognite.com/api/v1/#operation/byIdsAssets>`_
 
         Args:
             ids (List[int], optional): IDs
             external_ids (List[str], optional): External IDs
+            ignore_unknown_ids (bool): Ignore IDs and external IDs that are not found rather than throw an exception.
 
         Returns:
             AssetList: The requested assets.
@@ -137,18 +153,23 @@ class AssetsAPI(APIClient):
 
                 >>> from cognite.client import CogniteClient
                 >>> c = CogniteClient()
-                >>> res = c.assets.retrieve_multiple(external_ids=["abc", "def"])
+                >>> res = c.assets.retrieve_multiple(external_ids=["abc", "def"], ignore_unknown_ids=True)
         """
         utils._auxiliary.assert_type(ids, "id", [List], allow_none=True)
         utils._auxiliary.assert_type(external_ids, "external_id", [List], allow_none=True)
-        return self._retrieve_multiple(ids=ids, external_ids=external_ids, wrap_ids=True)
+        return self._retrieve_multiple(
+            ids=ids, external_ids=external_ids, ignore_unknown_ids=ignore_unknown_ids, wrap_ids=True
+        )
 
     def list(
         self,
         name: str = None,
         parent_ids: List[int] = None,
+        parent_external_ids: List[str] = None,
         root_ids: List[int] = None,
         root_external_ids: List[str] = None,
+        asset_subtree_ids: List[int] = None,
+        asset_subtree_external_ids: List[str] = None,
         metadata: Dict[str, str] = None,
         source: str = None,
         created_time: Union[Dict[str, Any], TimestampRange] = None,
@@ -164,9 +185,12 @@ class AssetsAPI(APIClient):
 
         Args:
             name (str): Name of asset. Often referred to as tag.
-            parent_ids (List[int]): List of parent ids to filter on.
+            parent_ids (List[int]): Return only the direct descendants of the specified assets.
+            parent_external_ids (List[str]): Return only the direct descendants of the specified assets.
             root_ids (List[int], optional): List of root ids ids to filter on.
             root_external_ids (List[str], optional): List of root external ids to filter on.
+            asset_subtree_ids (List[int]): List of asset subtrees ids to filter on.
+            asset_subtree_external_ids (List[str]): List of asset subtrees external ids to filter on.
             metadata (Dict[str, str]): Custom, application specific metadata. String key -> String value.
             source (str): The source of this asset.
             created_time (Union[Dict[str, Any], TimestampRange]): Range between two timestamps.
@@ -210,11 +234,15 @@ class AssetsAPI(APIClient):
         # dict option for backward compatibility
         if (root_ids and not isinstance(root_ids[0], dict)) or root_external_ids:
             root_ids = self._process_ids(root_ids, root_external_ids, wrap_ids=True)
+        if asset_subtree_ids or asset_subtree_external_ids:
+            asset_subtree_ids = self._process_ids(asset_subtree_ids, asset_subtree_external_ids, wrap_ids=True)
 
         filter = AssetFilter(
             name=name,
             parent_ids=parent_ids,
+            parent_external_ids=parent_external_ids,
             root_ids=root_ids,
+            asset_subtree_ids=asset_subtree_ids,
             metadata=metadata,
             source=source,
             created_time=created_time,
@@ -281,7 +309,11 @@ class AssetsAPI(APIClient):
         return _AssetPoster(assets, client=self).post()
 
     def delete(
-        self, id: Union[int, List[int]] = None, external_id: Union[str, List[str]] = None, recursive: bool = False
+        self,
+        id: Union[int, List[int]] = None,
+        external_id: Union[str, List[str]] = None,
+        recursive: bool = False,
+        ignore_unknown_ids: bool = False,
     ) -> None:
         """`Delete one or more assets <https://doc.cognitedata.com/api/v1/#operation/deleteAssets>`_
 
@@ -289,6 +321,7 @@ class AssetsAPI(APIClient):
             id (Union[int, List[int]): Id or list of ids
             external_id (Union[str, List[str]]): External ID or list of exgernal ids
             recursive (bool): Recursively delete whole asset subtrees under given ids. Defaults to False.
+            ignore_unknown_ids (bool): Ignore IDs and external IDs that are not found rather than throw an exception.
 
         Returns:
             None
@@ -302,7 +335,10 @@ class AssetsAPI(APIClient):
                 >>> c.assets.delete(id=[1,2,3], external_id="3")
         """
         self._delete_multiple(
-            ids=id, external_ids=external_id, wrap_ids=True, extra_body_fields={"recursive": recursive}
+            ids=id,
+            external_ids=external_id,
+            wrap_ids=True,
+            extra_body_fields={"recursive": recursive, "ignoreUnknownIds": ignore_unknown_ids},
         )
 
     def update(self, item: Union[Asset, AssetUpdate, List[Union[Asset, AssetUpdate]]]) -> Union[Asset, AssetList]:
@@ -335,13 +371,19 @@ class AssetsAPI(APIClient):
         return self._update_multiple(items=item)
 
     def search(
-        self, name: str = None, description: str = None, filter: Union[AssetFilter, Dict] = None, limit: int = 100
+        self,
+        name: str = None,
+        description: str = None,
+        query: str = None,
+        filter: Union[AssetFilter, Dict] = None,
+        limit: int = 100,
     ) -> AssetList:
         """`Search for assets <https://docs.cognite.com/api/v1/#operation/searchAssets>`_
 
         Args:
             name (str): Fuzzy match on name.
             description (str): Fuzzy match on description.
+            query (str): Whitespace-separated terms to search for in assets. Does a best-effort fuzzy search in relevant fields (currently name and description) for variations of any of the search terms, and orders results by relevance.
             filter (Union[AssetFilter, Dict]): Filter to apply. Performs exact match on these fields.
             limit (int): Maximum number of results to return.
 
@@ -362,13 +404,21 @@ class AssetsAPI(APIClient):
                 >>> c = CogniteClient()
                 >>> res = c.assets.search(filter={"name": "some name"})
 
+            Search for assets by improved multi-field fuzzy search::
+
+                >>> from cognite.client import CogniteClient
+                >>> c = CogniteClient()
+                >>> res = c.assets.search(query="TAG 30 XV")
+
             Search for assets using multiple filters, finding all assets with name similar to `xyz` with parent asset `123` or `456` with source `some source`::
 
                 >>> from cognite.client import CogniteClient
                 >>> c = CogniteClient()
                 >>> res = c.assets.search(name="xyz",filter={"parent_ids": [123,456],"source": "some source"})
         """
-        return self._search(search={"name": name, "description": description}, filter=filter, limit=limit)
+        return self._search(
+            search={"name": name, "description": description, "query": query}, filter=filter, limit=limit
+        )
 
     def retrieve_subtree(self, id: int = None, external_id: str = None, depth: int = None) -> AssetList:
         """Retrieve the subtree for this asset up to a specified depth.