+++ cognite/client/data_classes/events.py
@@ -66,7 +66,9 @@ class EventFilter(CogniteFilter):
         end_time (Union[Dict[str, Any], TimestampRange]): Range between two timestamps.
         metadata (Dict[str, str]): Custom, application specific metadata. String key -> String value. Limits: Maximum length of key is 32 bytes, value 512 bytes, up to 16 key-value pairs.
         asset_ids (List[int]): Asset IDs of related equipment that this event relates to.
-        root_asset_ids (List[Dict[str, Any]]): The IDs of the root assets that the related assets should be children of.
+        asset_external_ids (List[str]): Asset External IDs of related equipment that this event relates to.
+        root_asset_ids (List[Dict[str, Any]]): Only include events that have a related asset in a tree rooted at any of these root assetIds.
+        asset_subtree_ids (List[Dict[str, Any]]): Only include events that have a related asset in a subtree rooted at any of these assetIds.
         source (str): The source of this event.
         type (str): The event type
         subtype (str): The event subtype
@@ -82,7 +84,9 @@ class EventFilter(CogniteFilter):
         end_time: Union[Dict[str, Any], TimestampRange] = None,
         metadata: Dict[str, str] = None,
         asset_ids: List[int] = None,
+        asset_external_ids: List[str] = None,
         root_asset_ids: List[Dict[str, Any]] = None,
+        asset_subtree_ids: List[Dict[str, Any]] = None,
         source: str = None,
         type: str = None,
         subtype: str = None,
@@ -95,7 +99,9 @@ class EventFilter(CogniteFilter):
         self.end_time = end_time
         self.metadata = metadata
         self.asset_ids = asset_ids
+        self.asset_external_ids = asset_external_ids
         self.root_asset_ids = root_asset_ids
+        self.asset_subtree_ids = asset_subtree_ids
         self.source = source
         self.type = type
         self.subtype = subtype