+++ tests/tests_integration/test_api/test_model_hosting/test_models.py
@@ -21,7 +21,7 @@ class TestModels:
     def mock_online_predict_ok(self, rsps):
         rsps.add(
             rsps.POST,
-            MODELS_API._get_base_url_with_base_path() + "/analytics/models/model1/predict",
+            MODELS_API._get_base_url_with_base_path() + "/modelhosting/models/model1/predict",
             status=200,
             json={"predictions": [1, 2, 3]},
         )
@@ -31,7 +31,7 @@ class TestModels:
     def mock_online_predict_fail(self, rsps):
         rsps.add(
             rsps.POST,
-            MODELS_API._get_base_url_with_base_path() + "/analytics/models/model1/predict",
+            MODELS_API._get_base_url_with_base_path() + "/modelhosting/models/model1/predict",
             status=200,
             json={"error": {"message": "User error", "code": 200}},
         )