+++ tests/tests_unit/test_api/test_3d.py
@@ -43,6 +43,12 @@ class Test3DModels:
         assert isinstance(res, ThreeDModelList)
         assert mock_3d_model_response.calls[0].response.json()["items"] == res.dump(camel_case=True)
 
+    def test_list_published_default(self, mock_3d_model_response):
+        res = THREE_D_API.models.list(limit=100)
+        assert isinstance(res, ThreeDModelList)
+        assert mock_3d_model_response.calls[0].response.json()["items"] == res.dump(camel_case=True)
+        assert "published" not in mock_3d_model_response.calls[0].request.path_url
+
     def test_update_with_update_object(self, mock_3d_model_response):
         update = ThreeDModelUpdate(id=1).name.set("bla")
         res = THREE_D_API.models.update(update)