+++ tests/tests_unit/test_api/test_signatures.py
@@ -10,8 +10,16 @@ class TestListAndIterSignatures:
     @pytest.mark.parametrize(
         "api, filter, ignore",
         [
-            (assets.AssetsAPI, assets.AssetFilter, ["root_external_ids", "aggregated_properties", "partitions"]),
-            (events.EventsAPI, events.EventFilter, ["root_asset_external_ids", "partitions", "sort"]),
+            (
+                assets.AssetsAPI,
+                assets.AssetFilter,
+                ["root_external_ids", "asset_subtree_external_ids", "aggregated_properties", "partitions"],
+            ),
+            (
+                events.EventsAPI,
+                events.EventFilter,
+                ["root_asset_external_ids", "asset_subtree_external_ids", "partitions", "sort"],
+            ),
             (files.FilesAPI, files.FileMetadataFilter, ["root_asset_external_ids"]),
             (sequences.SequencesAPI, sequences.SequenceFilter, []),
         ],